Change result type for VGA->RGB conversion

+++ Pre2/AssetConverter.cs
@@ -495,12 +495,12 @@ namespace Pre2
             return palettes;
         }
 
-        private static int ConvertVgaToRgb(int sixBitValue)
+        private static byte ConvertVgaToRgb(byte sixBitValue)
         {
             sixBitValue &= 0x3F; // make sure it's really 6-bit value
             // 6-bit VGA to 8-bit RGB, approximation to Round(sixBitValue * 255 / 63)
-            //int eightBitValue = (sixBitValue * 255) / 63; // quite bad
-            int eightBitValue = (sixBitValue << 2) | (sixBitValue >> 4); // almost correct
+            //byte eightBitValue = (byte) ((sixBitValue * 255) / 63); // quite bad
+            byte eightBitValue = (byte) ((sixBitValue << 2) | (sixBitValue >> 4)); // almost correct
             return eightBitValue;
         }
